agent/requirements.txt
livekit-agents[images]
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-google
livekit-plugins-noise-cancellation
livekit-plugins-turn-detector
livekit-plugins-bey
mem0ai
duckduckgo-search
langchain_community
requests
python-dotenv